+++ betty/assets/public/localized/index.html.j2
@@ -1,5 +1,5 @@
 {% extends 'base.html.j2' %}
-{% set page_title = _('Welcome to %(title)s') | format(title=app.configuration.title) %}
+{% set page_title = _('Welcome to %(title)s') | format(title=app.project.configuration.title) %}
 {% block page_content %}
     <p>
         {% trans %}
@@ -9,7 +9,7 @@
     {% if 'betty.demo.Demo' in app.extensions %}
         <p>
             {% set liberta_lankester_label -%}
-                {%- with person = app.ancestry.entities['Person']['betty-demo-liberta-lankester'] -%}
+                {%- with person = app.project.ancestry.entities['Person']['betty-demo-liberta-lankester'] -%}
                     {% include 'label/person.html.j2' %}
                 {%- endwith -%}
             {%- endset %}
@@ -28,7 +28,7 @@
     {% set template_override_template %}
         {% raw %}
             {% extends 'base.html.j2' %}
-            {% set page_title = _('Welcome to %(title)s') | format(title=app.configuration.title) %}
+            {% set page_title = _('Welcome to %(title)s') | format(title=app.project.configuration.title) %}
             {% block page_content %}
                 {# Remove this line, and replace it with your own front page content. #}
             {% endblock %}
