+++ betty/assets/public/static/index.html.j2
@@ -1,7 +1,7 @@
 <!doctype html>
 <html>
 <head>
-    <title>{{ app.configuration.title }}</title>
+    <title>{{ app.project.configuration.title }}</title>
     <meta http-equiv="refresh" content="0; url={{ '/index.html' | url }}">
 </head>
 <body>
