+++ betty/assets/public/static/sitemap.xml.j2
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.sitemaps.org/schemas/sitemap/0.9 http://www.sitemaps.org/schemas/sitemap/0.9/sitemap.xsd">
-    {% for locale_configuration in app.configuration.locales %}
+    {% for locale_configuration in app.project.configuration.locales %}
         {% enter app.activate_locale(locale_configuration.locale) %}
-            {% for entity in app.ancestry.entities | reject('generated_entity_id') %}
+            {% for entity in app.project.ancestry.entities | reject('generated_entity_id') %}
                 <url>
                     <loc>{{ entity | url(absolute=true) }}</loc>
                 </url>
