+++ betty/assets/templates/base.html.j2
@@ -4,15 +4,15 @@
 <html lang="{{ app.locale }}"
       prefix="foaf: http://xmlns.com/foaf/0.1/ og: http://ogp.me/ns# rel: http://purl.org/vocab/relationship/">
 <head>
-    <title>{% if page_title is defined %}{{ page_title | striptags }} - {% endif %}{{ app.configuration.title }}</title>
+    <title>{% if page_title is defined %}{{ page_title | striptags }} - {% endif %}{{ app.project.configuration.title }}</title>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <meta name="generator" content="Betty (https://github.com/bartfeenstra/betty)"/>
-    {% if app.configuration.author %}
-        <meta name="author" content="{{ app.configuration.author }}"/>
+    {% if app.project.configuration.author %}
+        <meta name="author" content="{{ app.project.configuration.author }}"/>
     {% endif %}
-    <meta name="og:title" content="{%- if page_title is defined -%}{{ page_title | striptags }}{%- else -%}{{ app.configuration.title }}{%- endif -%}"/>
-    <meta name="og:site_name" content="{{ app.configuration.title }}"/>
+    <meta name="og:title" content="{%- if page_title is defined -%}{{ page_title | striptags }}{%- else -%}{{ app.project.configuration.title }}{%- endif -%}"/>
+    <meta name="og:site_name" content="{{ app.project.configuration.title }}"/>
     {% if page_resource is defined and page_recource is has_files %}
         {% for file in page_resource.files %}
             {% if file.media_type.type == 'image' %}
@@ -46,8 +46,8 @@
 </head>
 {% if page_resource is has_files %}
     {% set page_image = page_resource.associated_files | selectattr('media_type.type', 'eq', 'image') | first %}
-{% elif app.configuration.theme.background_image_id and app.configuration.theme.background_image_id in app.ancestry.entities['File'] %}
-    {% set page_image = app.ancestry.entities['File'][app.configuration.theme.background_image_id] %}
+{% elif app.project.configuration.theme.background_image_id and app.project.configuration.theme.background_image_id in app.project.ancestry.entities['File'] %}
+    {% set page_image = app.project.ancestry.entities['File'][app.project.configuration.theme.background_image_id] %}
 {% endif %}
 <body>
 <script>
@@ -59,13 +59,13 @@
 {% endif %}
 <div id="page">
     <nav id="nav-primary">
-        <a id="site-title" href="{{ '/index.html' | url }}" title="{{ app.configuration.title }}">{{ app.configuration.title }}</a>
+        <a id="site-title" href="{{ '/index.html' | url }}" title="{{ app.project.configuration.title }}">{{ app.project.configuration.title }}</a>
         <div id="search" data-betty-search-index="{{ '/search-index.json' | url }}">
             <div class="overlay-controls">
                 <span class="overlay-control overlay-close" title="{% trans %}Exit the search{% endtrans %}">{% trans %}Exit the search{% endtrans %}</span>
             </div>
             <form>
-                {% set search_keywords_example_person_name = app.ancestry.entities['Person'] | map(attribute='name') | first | default(none) %}
+                {% set search_keywords_example_person_name = app.project.ancestry.entities['Person'] | map(attribute='name') | first | default(none) %}
                 {% if search_keywords_example_person_name %}
                     {% set search_keywords_example -%}
                         {% trans example = person_macros.name_label(search_keywords_example_person_name, embedded=True) | striptags -%}
@@ -91,28 +91,28 @@
             <h2 class="nav-primary-action">{% trans %}Menu{% endtrans %}</h2>
             <div class="nav-primary-expanded">
                 <ul class="nav-secondary">
-                    {% if app.ancestry.entities['Person'] | length > 0 %}
+                    {% if app.project.ancestry.entities['Person'] | length > 0 %}
                         <li><a href="{{ '/person/index.html' | url }}">{% trans %}People{% endtrans %}</a></li>
                     {% endif %}
-                    {% if app.ancestry.entities['Event'] | length > 0 %}
+                    {% if app.project.ancestry.entities['Event'] | length > 0 %}
                         <li><a href="{{ '/event/index.html' | url }}">{% trans %}Timeline{% endtrans %}</a></li>
                     {% endif %}
-                    {% if app.ancestry.entities['Place'] | length > 0 %}
+                    {% if app.project.ancestry.entities['Place'] | length > 0 %}
                         <li><a href="{{ '/place/index.html' | url }}">{% trans %}Places{% endtrans %}</a></li>
                     {% endif %}
-                    {% if app.ancestry.entities['Source'] | length > 0 %}
+                    {% if app.project.ancestry.entities['Source'] | length > 0 %}
                         <li><a href="{{ '/source/index.html' | url }}">{% trans %}Sources{% endtrans %}</a></li>
                     {% endif %}
                 </ul>
             </div>
         </div>
-        {% if page_resource is defined and app.configuration.multilingual %}
+        {% if page_resource is defined and app.project.configuration.multilingual %}
             <div id="nav-locale" class="nav-primary-expandable">
                 <h2 class="nav-primary-action">{% trans %}Language{% endtrans %}</h2>
                 <div class="nav-primary-expanded">
                     <ul class="nav-secondary">
                     {% set ns = namespace(available_locales=[]) %}
-                    {% for available_locale_configuration in app.configuration.locales %}
+                    {% for available_locale_configuration in app.project.configuration.locales %}
                         {% set available_locale_data = available_locale_configuration.alias | locale_get_data %}
                         {% do ns.available_locales.append((available_locale_configuration.locale, available_locale_data.get_display_name())) %}
                     {% endfor %}
