+++ betty/assets/templates/footer.html.j2
@@ -1,6 +1,6 @@
-{% if app.configuration.author %}
+{% if app.project.configuration.author %}
     <p>
-        {% trans author = app.configuration.author %}© Copyright {{ author }}, unless otherwise credited{% endtrans %}
+        {% trans author = app.project.configuration.author %}© Copyright {{ author }}, unless otherwise credited{% endtrans %}
     </p>
 {% endif %}
 <p>{% trans %}A family history as told by <a href="https://github.com/bartfeenstra/betty">Betty👵</a>{% endtrans %}</p>
