+++ betty/http_api_doc/__init__.py
@@ -21,7 +21,7 @@ class HttpApiDoc(Extension, Generator, GuiBuilder, NpmBuilder):
 
     async def generate(self) -> None:
         assets_directory_path = await self._app.extensions[_Npm].ensure_assets(self)
-        copy2(assets_directory_path / 'http-api-doc.js', self._app.configuration.www_directory_path / 'http-api-doc.js')
+        copy2(assets_directory_path / 'http-api-doc.js', self._app.project.configuration.www_directory_path / 'http-api-doc.js')
 
     @classmethod
     def assets_directory_path(cls) -> Optional[Path]:
