+++ betty/http_api_doc/assets/public/localized/api/index.html.j2
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 <head>
-    <title>{% trans %}API documentation{% endtrans %} - {{ app.configuration.title }}</title>
+    <title>{% trans %}API documentation{% endtrans %} - {{ app.project.configuration.title }}</title>
     <meta charset="utf-8"/>
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link href="https://fonts.googleapis.com/css?family=Montserrat:300,400,700|Roboto:300,400,700" rel="stylesheet">
