+++ betty/maps/assets/betty.npm._Npm/src/configuration.json.j2
@@ -1,3 +1,3 @@
 {{ {
-    'rootPath': app.configuration.root_path,
+    'rootPath': app.project.configuration.root_path,
 } | json }}
