+++ betty/tests/http_api_doc/test__init__.py
@@ -1,7 +1,8 @@
 from betty.asyncio import sync
 from betty.http_api_doc import HttpApiDoc
 from betty.generate import generate
-from betty.app import App, AppExtensionConfiguration
+from betty.app import App
+from betty.project import ProjectExtensionConfiguration
 from betty.tests import patch_cache, TestCase
 
 
@@ -10,7 +11,7 @@ class HttpApiDocTest(TestCase):
     @sync
     async def test(self):
         async with App() as app:
-            app.configuration.extensions.add(AppExtensionConfiguration(HttpApiDoc))
+            app.project.configuration.extensions.add(ProjectExtensionConfiguration(HttpApiDoc))
             await generate(app)
-            self.assertTrue((app.configuration.output_directory_path / 'www' / 'api' / 'index.html').is_file())
-            self.assertTrue((app.configuration.output_directory_path / 'www' / 'http-api-doc.js').is_file())
+            self.assertTrue((app.project.configuration.output_directory_path / 'www' / 'api' / 'index.html').is_file())
+            self.assertTrue((app.project.configuration.output_directory_path / 'www' / 'http-api-doc.js').is_file())
