+++ betty/tests/test_generate.py
@@ -8,17 +8,18 @@ import html5lib
 from lxml import etree
 
 from betty import json
-from betty.app import App, LocaleConfiguration
+from betty.app import App
 from betty.asyncio import sync
 from betty.generate import generate
 from betty.model.ancestry import Person, Place, Source, PlaceName, File, Event, Citation
 from betty.model.event_type import Birth
+from betty.project import LocaleConfiguration
 from betty.tests import TestCase
 
 
 class GenerateTestCase(TestCase):
     def assert_betty_html(self, app: App, url_path: str) -> Path:
-        file_path = app.configuration.www_directory_path / Path(url_path.lstrip('/'))
+        file_path = app.project.configuration.www_directory_path / Path(url_path.lstrip('/'))
         self.assertTrue(file_path.exists(), '%s does not exist' % file_path)
         with open(file_path) as f:
             parser = html5lib.HTMLParser(strict=True)
@@ -26,7 +27,7 @@ class GenerateTestCase(TestCase):
         return file_path
 
     def assert_betty_json(self, app: App, url_path: str, schema_definition: str) -> Path:
-        file_path = app.configuration.www_directory_path / Path(url_path.lstrip('/'))
+        file_path = app.project.configuration.www_directory_path / Path(url_path.lstrip('/'))
         self.assertTrue(file_path.exists(), '%s does not exist' % file_path)
         with open(file_path) as f:
             json.validate(stdjson.load(f), schema_definition, app)
@@ -52,7 +53,7 @@ class GenerateTest(GenerateTestCase):
         async with App() as app:
             with NamedTemporaryFile() as f:
                 file = File('FILE1', Path(f.name))
-                app.ancestry.entities.append(file)
+                app.project.ancestry.entities.append(file)
                 await generate(app)
             self.assert_betty_html(app, '/file/%s/index.html' % file.id)
             self.assert_betty_json(app, '/file/%s/index.json' % file.id, 'file')
@@ -68,7 +69,7 @@ class GenerateTest(GenerateTestCase):
     async def test_place(self):
         async with App() as app:
             place = Place('PLACE1', [PlaceName('one')])
-            app.ancestry.entities.append(place)
+            app.project.ancestry.entities.append(place)
             await generate(app)
         self.assert_betty_html(app, '/place/%s/index.html' % place.id)
         self.assert_betty_json(app, '/place/%s/index.json' % place.id, 'place')
@@ -84,7 +85,7 @@ class GenerateTest(GenerateTestCase):
     async def test_person(self):
         async with App() as app:
             person = Person('PERSON1')
-            app.ancestry.entities.append(person)
+            app.project.ancestry.entities.append(person)
             await generate(app)
         self.assert_betty_html(app, '/person/%s/index.html' % person.id)
         self.assert_betty_json(app, '/person/%s/index.json' % person.id, 'person')
@@ -100,7 +101,7 @@ class GenerateTest(GenerateTestCase):
     async def test_event(self):
         async with App() as app:
             event = Event('EVENT1', Birth())
-            app.ancestry.entities.append(event)
+            app.project.ancestry.entities.append(event)
             await generate(app)
         self.assert_betty_html(app, '/event/%s/index.html' % event.id)
         self.assert_betty_json(app, '/event/%s/index.json' % event.id, 'event')
@@ -109,7 +110,7 @@ class GenerateTest(GenerateTestCase):
     async def test_citation(self):
         async with App() as app:
             citation = Citation('CITATION1', Source('A Little Birdie'))
-            app.ancestry.entities.append(citation)
+            app.project.ancestry.entities.append(citation)
             await generate(app)
         self.assert_betty_html(app, '/citation/%s/index.html' % citation.id)
         self.assert_betty_json(app, '/citation/%s/index.json' %
@@ -126,7 +127,7 @@ class GenerateTest(GenerateTestCase):
     async def test_source(self):
         async with App() as app:
             source = Source('SOURCE1', 'A Little Birdie')
-            app.ancestry.entities.append(source)
+            app.project.ancestry.entities.append(source)
             await generate(app)
         self.assert_betty_html(app, '/source/%s/index.html' % source.id)
         self.assert_betty_json(app, '/source/%s/index.json' % source.id, 'source')
@@ -136,7 +137,7 @@ class MultilingualTest(GenerateTestCase):
     @sync
     async def test_root_redirect(self):
         app = App()
-        app.configuration.locales.replace([
+        app.project.configuration.locales.replace([
             LocaleConfiguration('nl'),
             LocaleConfiguration('en'),
         ])
@@ -149,7 +150,7 @@ class MultilingualTest(GenerateTestCase):
     @sync
     async def test_public_localized_resource(self):
         app = App()
-        app.configuration.locales.replace([
+        app.project.configuration.locales.replace([
             LocaleConfiguration('nl'),
             LocaleConfiguration('en'),
         ])
@@ -165,13 +166,13 @@ class MultilingualTest(GenerateTestCase):
     @sync
     async def test_entity(self):
         app = App()
-        app.configuration.locales.replace([
+        app.project.configuration.locales.replace([
             LocaleConfiguration('nl'),
             LocaleConfiguration('en'),
         ])
         async with app:
             person = Person('PERSON1')
-            app.ancestry.entities.append(person)
+            app.project.ancestry.entities.append(person)
             await generate(app)
         with open(self.assert_betty_html(app, '/nl/person/%s/index.html' % person.id)) as f:
             translation_link = '<a href="/en/person/%s/index.html" hreflang="en" lang="en" rel="alternate">English</a>' % person.id
@@ -191,9 +192,9 @@ class ResourceOverrideTest(GenerateTestCase):
             with open(str(localized_assets_directory_path / 'index.html.j2'), 'w') as f:
                 f.write('{% block page_content %}Betty was here{% endblock %}')
             async with App() as app:
-                app.configuration.assets_directory_path = assets_directory_path
+                app.project.configuration.assets_directory_path = assets_directory_path
                 await generate(app)
-        with open(app.configuration.www_directory_path / 'index.html') as f:
+        with open(app.project.configuration.www_directory_path / 'index.html') as f:
             self.assertIn('Betty was here', f.read())
 
 
@@ -206,6 +207,6 @@ class SitemapGenerateTest(GenerateTestCase):
         with open(Path(__file__).parent / 'test_generate_assets' / 'sitemap.xsd') as f:
             schema_doc = etree.parse(f)
         schema = etree.XMLSchema(schema_doc)
-        with open(app.configuration.www_directory_path / 'sitemap.xml') as f:
+        with open(app.project.configuration.www_directory_path / 'sitemap.xml') as f:
             sitemap_doc = etree.parse(f)
         schema.validate(sitemap_doc)
