+++ betty/tests/trees/test__init__.py
@@ -1,7 +1,8 @@
 from betty.asyncio import sync
 from betty.generate import generate
+from betty.project import ProjectExtensionConfiguration
 from betty.trees import Trees
-from betty.app import App, AppExtensionConfiguration
+from betty.app import App
 from betty.tests import patch_cache, TestCase
 
 
@@ -10,12 +11,12 @@ class TreesTest(TestCase):
     @sync
     async def test_post_generate_event(self):
         async with App() as app:
-            app.configuration.debug = True
-            app.configuration.extensions.add(AppExtensionConfiguration(Trees))
+            app.project.configuration.debug = True
+            app.project.configuration.extensions.add(ProjectExtensionConfiguration(Trees))
             await generate(app)
-        with open(app.configuration.www_directory_path / 'trees.js', encoding='utf-8') as f:
+        with open(app.project.configuration.www_directory_path / 'trees.js', encoding='utf-8') as f:
             betty_js = f.read()
         self.assertIn('trees.js', betty_js)
-        with open(app.configuration.www_directory_path / 'trees.css', encoding='utf-8') as f:
+        with open(app.project.configuration.www_directory_path / 'trees.css', encoding='utf-8') as f:
             betty_css = f.read()
         self.assertIn('.tree', betty_css)
