+++ betty/trees/__init__.py
@@ -28,7 +28,7 @@ class Trees(Extension, CssProvider, JsProvider, Generator, GuiBuilder, NpmBuilde
 
     async def generate(self) -> None:
         assets_directory_path = await self._app.extensions[_Npm].ensure_assets(self)
-        self._copy_npm_build(assets_directory_path, self._app.configuration.www_directory_path)
+        self._copy_npm_build(assets_directory_path, self._app.project.configuration.www_directory_path)
 
     @classmethod
     def assets_directory_path(cls) -> Optional[Path]:
